(Request: Adrian) File restructuring
