--- client/src/pages/SoilGame/TESTING/testGameHelper.ts
@@ -1,2 +1,0 @@
-export * from '../utils/GameHelper_REP';
-
